Separate fixtures for WalletService and TransactionService
